hugo: cleanup utils
