+++ eagle/utils.go
@@ -47,3 +47,22 @@ func cleanID(id string) string {
 	id = strings.TrimPrefix(id, "/")
 	return "/" + id
 }
+
+// Borrowed from https://github.com/jlelse/GoBlog/blob/master/utils.go
+func Slugify(str string) string {
+	return strings.Map(func(c rune) rune {
+		if c >= 'a' && c <= 'z' || c >= '0' && c <= '9' {
+			// Is lower case ASCII or number, return unmodified
+			return c
+		} else if c >= 'A' && c <= 'Z' {
+			// Is upper case ASCII, make lower case
+			return c + 'a' - 'A'
+		} else if c == ' ' || c == '-' || c == '_' {
+			// Space, replace with '-'
+			return '-'
+		} else {
+			// Drop character
+			return -1
+		}
+	}, str)
+}
